main_fanc/src/import_gcs_schema/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery>=2.18.0,<3.0.0
google-cloud-storage>=2.8.0,<3.0.0